--- a/perf-event-open-sys/src/lib.rs
+++ b/perf-event-open-sys/src/lib.rs
@@ -182,6 +182,10 @@
 #[path = "bindings_aarch64.rs"]
 pub mod bindings;
 
+#[cfg(target_arch = "arm")]
+#[path = "bindings_armv7l.rs"]
+pub mod bindings;
+
 #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
 #[path = "bindings_x86_64.rs"]
 pub mod bindings;