This adds armv7 bindings. _(Really just generated and added a couple of lines to existing sources)._

Generated binding on raspi-os

    Linux raspberrypi 5.15.84-v7+ #1613 SMP Thu Jan 5 11:59:48 GMT 2023 armv7l GNU/Linux

Tested on the same system, but cross compiled with cross-rs and musl.

cargo test results:

```
$ cargo test
   Compiling bitflags v1.3.2
   Compiling perf-event-open-sys v4.0.0 (/home/m/src/perf-event/perf-event-open-sys)
   Compiling perf-event v0.4.8 (/home/m/src/perf-event/perf-event)
    Finished test [unoptimized + debuginfo] target(s) in 31.55s
     Running unittests src/lib.rs (target/debug/deps/perf_event-1542f7d8622ec449)

running 2 tests
test test_error_code_is_correct ... ok
test simple_build ... ok

test result: ok. 2 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.04s

     Running tests/breakpoint.rs (target/debug/deps/breakpoint-a90b7f46d62bb388)

running 2 tests
test data ... ok
test execute ... ok

test result: ok. 2 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.01s

     Running unittests src/lib.rs (target/debug/deps/perf_event_open_sys-bef037ddf6f30e62)

running 18 tests
test bindings::bindgen_test_layout___kernel_fd_set ... ok
test bindings::bindgen_test_layout___kernel_fsid_t ... ok
test bindings::bindgen_test_layout_perf_event_attr ... ok
test bindings::bindgen_test_layout_perf_event_attr__bindgen_ty_2 ... ok
test bindings::bindgen_test_layout_perf_event_attr__bindgen_ty_3 ... ok
test bindings::bindgen_test_layout_perf_event_attr__bindgen_ty_4 ... ok
test bindings::bindgen_test_layout_perf_branch_entry ... ok
test bindings::bindgen_test_layout_perf_event_attr__bindgen_ty_1 ... ok
test bindings::bindgen_test_layout_perf_event_header ... ok
test bindings::bindgen_test_layout_perf_event_mmap_page ... ok
test bindings::bindgen_test_layout_perf_event_mmap_page__bindgen_ty_1 ... ok
test bindings::bindgen_test_layout_perf_event_mmap_page__bindgen_ty_1__bindgen_ty_1 ... ok
test bindings::bindgen_test_layout_perf_event_query_bpf ... ok
test bindings::bindgen_test_layout_perf_mem_data_src ... ok
test bindings::bindgen_test_layout_perf_mem_data_src__bindgen_ty_1 ... ok
test bindings::bindgen_test_layout_perf_ns_link_info ... ok
test bindings::bindgen_test_layout_perf_sample_weight ... ok
test bindings::bindgen_test_layout_perf_sample_weight__bindgen_ty_1 ... ok

test result: ok. 18 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.01s

   Doc-tests perf-event

running 20 tests
test src/hooks.rs - hooks::RealHooks (line 194) ... ok
test src/events.rs - events::Cache (line 213) ... ok
test src/events.rs - events::Breakpoint (line 365) ... ok
test src/events.rs - events::Breakpoint (line 392) ... ok
test src/lib.rs - Builder (line 161) ... ok
test src/lib.rs - Builder (line 169) ... ok
test src/lib.rs - (line 10) ... ok
test src/lib.rs - Builder (line 180) ... ok
test src/lib.rs - Builder<'a>::kind (line 590) ... ok
test src/lib.rs - Counter::read_count_and_time (line 755) ... ok
test src/lib.rs - Counter (line 110) ... ok
test src/lib.rs - Counts (line 351) ... ok
test src/lib.rs - Counts (line 365) ... ok
test src/lib.rs - Counts::get (line 1043) ... ok
test src/lib.rs - Group (line 301) ... ignored
test src/lib.rs - Group (line 307) ... ignored
test src/lib.rs - Group::read (line 894) ... ignored
test src/lib.rs - Counts (line 382) ... ok
test src/lib.rs - Counts::iter (line 1058) ... ok
test src/lib.rs - Group (line 241) ... ok

test result: ok. 17 passed; 0 failed; 3 ignored; 0 measured; 0 filtered out; finished in 23.64s

   Doc-tests perf-event-open-sys

running 1 test
test src/lib.rs - (line 31) ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 2.93s
```